refactor: pull out AST operations in an analysis module
